# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClinicAI SaaS multi-tenant WhatsApp automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/handoffs:
    get:
      operationId: listHandoffs
      tags: [handoffs]
      summary: Lista handoffs ativos para a clínica
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de handoffs ativos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Handoff"
    post:
      operationId: createHandoff
      tags: [handoffs]
      summary: Atendente assume uma conversa (pausa a IA)
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHandoffBody"
      responses:
        "201":
          description: Handoff criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Handoff"
        "409":
          description: Já existe handoff ativo para esse telefone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clinics/{clinicId}/handoffs/{phone}:
    delete:
      operationId: deleteHandoff
      tags: [handoffs]
      summary: Encerra um handoff ativo
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Handoff encerrado
        "404":
          description: Handoff não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clinics/{clinicId}/handoffs/{phone}/messages:
    get:
      operationId: listHandoffMessages
      tags: [handoffs]
      summary: Histórico mesclado de mensagens (IA + atendente) para um paciente
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: phone
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Lista de mensagens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationHistoryItem"
    post:
      operationId: sendHandoffMessage
      tags: [handoffs]
      summary: Atendente envia mensagem pelo WhatsApp
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: integer
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendHandoffMessageBody"
      responses:
        "201":
          description: Mensagem enviada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationHistoryItem"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Handoff:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        patientPhone:
          type: string
        attendantId:
          type: integer
          nullable: true
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, clinicId, patientPhone, startedAt]
    CreateHandoffBody:
      type: object
      properties:
        patientPhone:
          type: string
      required: [patientPhone]
    ConversationHistoryItem:
      type: object
      properties:
        id:
          type: integer
        source:
          type: string
          enum: [ai, attendant, patient]
          description: "ai = resposta da IA; attendant = resposta do atendente humano; patient = mensagem do paciente"
        content:
          type: string
        patientPhone:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, source, content, patientPhone, createdAt]
    SendHandoffMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
...